memperbaiki classroom dan course sesuai school

+++ app/api/course/[id]/route.ts
@@ -1,4 +1,6 @@
+import { authOptions } from "@/lib/auth"
 import { prisma } from "@/lib/prisma"
+import { getServerSession } from "next-auth"
 
 import { NextResponse } from "next/server"
 
@@ -51,24 +53,41 @@ export async function GET(request: Request) {
     }
 }
 
+export async function DELETE(
+    request: Request
+) {
+    const session = await getServerSession(authOptions)
 
-export async function DELETE(request: Request) {
-    // const session = await getServerSession(authOptions)
+    if (!session || !session.user) {
+        return NextResponse.json({ message: "Unauthorized" }, { status: 401 })
+    }
 
-    // if (!session || !session.user) {
-    //     return new Response(
-    //         JSON.stringify({ message: "Unauthorized" }),
-    //         { status: 401 }
-    //     )
-    // }
+    const url = new URL(request.url)
+    const idString = url.pathname.split("/").pop()
 
-    const body = await request.json()
+    const id = parseInt(idString || "")
 
-    const deletedCourse = await prisma.course.delete({
-        where: {
-            id: body.id,
-        }
-    })
 
-    return NextResponse.json(deletedCourse, { status: 201 })
+    if (isNaN(id)) {
+        return new Response(
+            JSON.stringify({ message: "Invalid ID" }),
+            { status: 400 }
+        )
+    }
+    try {
+        await prisma.course.delete({
+            where: { id },
+        })
+
+        return NextResponse.json(
+            { message: "Course deleted successfully" },
+            { status: 200 }
+        )
+    } catch (error) {
+        console.error(error)
+        return NextResponse.json(
+            { error: "Failed to delete Course" },
+            { status: 500 }
+        )
+    }
 }
